chore(ssot): phase 9 - remove backend models shim and relocate notification intent
- Remove backend/api/src/models/Ad.ts compatibility shim and delete empty directory
- Update backend migrations and scripts to import canonical models from @esparex/core/models
- Relocate core/src/domain/NotificationIntent.ts to core/src/domains/notifications/domain/
- Export NotificationIntent from notifications domain barrel and update all consumers
- Update compatibility-marker-baseline.json to reflect removed shim
- All 18 repository gate checks pass with 100% health score
