+++ backend/api/src/controllers/admin/adminNotificationController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import mongoose from "mongoose";
 import { NOTIFICATION_TYPE } from "@esparex/contracts";
-import { NotificationIntent } from "@esparex/core/domain/NotificationIntent";
+import { NotificationIntent } from "@esparex/core/domains/notifications";
 import {
     createNotificationLog,
     createScheduledNotification,
