--- backend/api/src/models/Ad.ts
@@ -1,6 +1,0 @@
-// backend/src/models/Ad.ts
-// COMPATIBILITY SHIM: Re-exports the canonical Ad model from @core.
-// This file exists only for legacy CI guard compatibility and should not contain business logic.
-
-export * from "@esparex/core/models/Ad";
-export { default } from "@esparex/core/models/Ad";
