+++ core/package.json
@@ -40,7 +40,7 @@
     "./queues/*": "./dist/queues/*.js",
     "./jobs/*": "./dist/jobs/*.js",
     "./db/*": "./dist/db/*.js",
-    "./domain/NotificationIntent": "./dist/domain/NotificationIntent.js",
+    "./domains/notifications/domain/NotificationIntent": "./dist/domains/notifications/domain/NotificationIntent.js",
     "./lib/redis": "./dist/lib/redis.js",
     "./lib/location/formatLocation": "./dist/lib/location/formatLocation.js",
     "./migrations/*": "./dist/migrations/*.js"
