+++ core/src/__tests__/services/NotificationQueue.spec.ts
@@ -4,7 +4,7 @@
 
 import { Worker } from 'bullmq';
 import { NotificationDispatcher } from '../../services/notification/NotificationDispatcher';
-import { NotificationIntent } from '../../domain/NotificationIntent';
+import { NotificationIntent } from '../../domains/notifications';
 import { enqueueDeadLetter } from '../../queues/deadLetterQueue';
 import { notificationDeliveryProcessor } from '../../workers/notificationDeliveryWorker';
