+++ core/src/domains/notifications/index.ts
@@ -3,3 +3,4 @@ export * from './application/NotificationDispatcher';
 export * from './application/AdminNotificationService';
 export * from './application/SmartAlertService';
 export * from './application/PlansWalletNotificationService';
+export * from './domain/NotificationIntent';
