+++ core/src/services/SchedulerService.ts
@@ -3,7 +3,7 @@ import type { Types } from "mongoose";
 import ScheduledNotification from "../models/ScheduledNotification";
 import NotificationLog from "../models/NotificationLog";
 import { NOTIFICATION_TYPE } from '@esparex/contracts';
-import { NotificationIntent } from "../domain/NotificationIntent";
+import { NotificationIntent } from "../domains/notifications";
 import { NotificationDispatcher } from "./notification/NotificationDispatcher";
 import logger from "../utils/logger";
 import { runWithDistributedJobLock } from "../utils/distributedJobLock";
